+++ PopupBridge/src/main/java/com/braintreepayments/popupbridge/PopupBridge.java
@@ -8,16 +8,17 @@ import android.os.Bundle;
 import android.webkit.JavascriptInterface;
 import android.webkit.WebView;
 
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
+
 import com.braintreepayments.browserswitch.BrowserSwitchFragment;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.Set;
-
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentActivity;
-import androidx.fragment.app.FragmentManager;
 
 public class PopupBridge extends BrowserSwitchFragment {
 
@@ -31,6 +32,22 @@ public class PopupBridge extends BrowserSwitchFragment {
     private PopupBridgeMessageListener mMessageListener;
 
     public PopupBridge() {}
+
+    /**
+     * NEXT_MAJOR_VERSION: remove this method in favor of PopupBridge#newInstance(FragmentActivity, WebView)
+     * Create a new instance of {@link PopupBridge} and add it to the {@link AppCompatActivity}'s {@link android.support.v4.app.FragmentManager}.
+     *
+     * This will enable JavaScript in your WebView.
+     *
+     * @param activity The {@link AppCompatActivity} to add the {@link Fragment} to.
+     * @param webView The {@link WebView} to enable for PopupBridge.
+     * @return {@link PopupBridge}
+     * @throws IllegalArgumentException If the activity is not valid or the fragment cannot be added.
+     */
+    @SuppressLint("SetJavaScriptEnabled")
+    public static PopupBridge newInstance(AppCompatActivity activity, WebView webView) throws IllegalArgumentException {
+        return newInstance((FragmentActivity)activity, webView);
+    }
 
     /**
      * Create a new instance of {@link PopupBridge} and add it to the {@link FragmentActivity}'s {@link android.support.v4.app.FragmentManager}.